--- a/lib/logger.js
+++ b/lib/logger.js
@@ -1,19 +1,65 @@
 'use strict';
 
 const Hoek = require('@hapi/hoek');
-const moment = require('moment');
 const printf = require('util').format;
-const stringify = require('json-stringify-safe');
+const stringify = require('fast-safe-stringify');
 const inspect = require('util').inspect;
+const SonicBoom = require('sonic-boom');
 
 const internals = {
 	defaults: {
 		jsonOutput: true,
-		format: 'YYYY-MM-DD HH:mm:ss.SSS',
-		utc: false,
-		requestInfoFilter: (requestInfo) => requestInfo
+		formatTimestamp,
+		requestInfoFilter: (requestInfo) => requestInfo,
+		sonic: {fd: process.stdout.fd, sync: true}
 	}
 };
+
+function formatTimestamp(timestamp) {
+	const d = new Date(parseInt(timestamp, 10));
+
+	let month = d.getMonth() + 1;
+	month = month < 10 ? `0${month}` : month;
+
+	let day = d.getDate();
+	day = day < 10 ? `0${day}` : day;
+
+	let hours = d.getDate();
+	hours = hours < 10 ? `0${hours}` : hours;
+
+	let minutes = d.getMinutes();
+	minutes = minutes < 10 ? `0${minutes}` : minutes;
+
+	let seconds = d.getSeconds();
+	seconds = seconds < 10 ? `0${seconds}` : seconds;
+
+	return `${d.getFullYear()}-${month}-${day} ${hours}:${minutes}:${seconds}.${d.getMilliseconds()}`;
+}
+
+const noop = () => {};
+
+// function copied from https://github.com/pinojs/pino
+function buildSafeSonicBoom(opts) {
+	const stream = new SonicBoom(opts);
+	stream.on('error', filterBrokenPipe);
+	return stream;
+
+	function filterBrokenPipe(err) {
+		if (err.code === 'EPIPE') {
+			// If we get EPIPE, we should stop logging here
+			// however we have no control to the consumer of
+			// SonicBoom, so we just overwrite the write method
+			stream.write = noop;
+			stream.end = noop;
+			stream.flushSync = noop; // eslint-disable-line no-sync
+			stream.destroy = noop;
+			return;
+		}
+
+		stream.removeListener('error', filterBrokenPipe);
+		stream.emit('error', err);
+	}
+}
 
 /**
  * Default meta data to append to logs
@@ -25,6 +71,27 @@ function defaultMeta(request) {
 		return;
 	}
 	return {requestId: request.info.id};
+}
+
+/**
+ * defaultHandler for outputing the logs
+ * @param {Object} [opts] sonic boom options
+ * @return {Object}
+ */
+function defaultHandler(opts) {
+	const sonic = buildSafeSonicBoom(opts);
+	return {
+		log(str) {
+			sonic.write(str + '\n');
+		},
+
+		flush(sync = false) {
+			if (sync) {
+				return sonic.flushSync(); // eslint-disable-line no-sync
+			}
+			return sonic.flush();
+		}
+	};
 }
 
 /**
@@ -57,16 +124,17 @@ class Logger {
 	 *
 	 * @class Logger
 	 * @param {Object} [options]
-	 * @param {String} [options.format] date format string
-	 * @param {Boolean} [options.utc] UTC date
+	 * @param {Function} [options.formatTimestamp] function to format timestamp
 	 * @param {Boolean} [options.jsonOutput] output all data as stringified json
 	 * @param {Object} [options.handler] Handler implementing `log(message)` method, defaults to console logging
+	 * @param {Object} [options.sonic] sonic boom options if using default handler
 	 * @param {Function} [options.meta] Should return meta data as an Object that is appended to logs. Default returns {requestId:request.info.id} This function doesnt always recieve request object.
 	 */
 	constructor(options) {
 		this.opts = Hoek.applyToDefaults(internals.defaults, options || {});
-		this.handler = this.opts.handler || console;
+		this.handler = this.opts.handler || defaultHandler(this.opts.sonic);
 		this.meta = this.opts.meta || defaultMeta;
+		this.formatTime = this.opts.formatTimestamp;
 	}
 
 	/**
@@ -81,20 +149,6 @@ class Logger {
 			data: this.meta(request),
 			log: printf('%s, stack:\n%s', data.error.message, data.error.stack)
 		});
-	}
-
-	/**
-	 * Format timestamp
-	 * @param  {Integer} timestamp
-	 * @return {String}
-	 */
-	formatTime(timestamp) {
-		const m = moment(parseInt(timestamp, 10));
-		if (!this.opts.utc) {
-			m.local();
-		}
-
-		return m.format(this.opts.format);
 	}
 
 	/**
@@ -218,19 +272,18 @@ class Logger {
 			return this.humanReadableFormatter(obj);
 		}
 
-		let formatted = {
+		const formatted = {
 			_time: this.formatTime(obj.timestamp),
 			_tags: obj.tags
 		};
 
-		formatted = Hoek.applyToDefaults(formatted, obj.data || {});
-		formatted = Hoek.applyToDefaults(formatted, obj.requestInfo || {});
+		Object.assign(formatted, obj.data || {}, obj.requestInfo || {});
 
 		if (obj.log && !Array.isArray(obj.log) && typeof obj.log === 'object') {
 			if (obj.log instanceof Error) {
 				formatted.error = obj.log.stack;
 			} else {
-				formatted = Hoek.applyToDefaults(formatted, obj.log);
+				Object.assign(formatted, obj.log);
 			}
 		} else {
 			formatted.msg = obj.log;