Was reading about performance and logging and thought it would be interesting to see how hapi-log stands compared to others. Simply put it was horrible.

This introduces some breaking changes. Date formatting is just a sad thing, this has been replaced with a simple one that conforms to our most common style with a new option `formatTimestamp`  to pass in your own function for formatting a timestamp. All old formatting options for times has been removed.

### Benchmarks

The benchmarks are writing 10 lines 10000 times.

**Before**

```
benchString*10000: 2373.986ms
benchError*10000: 2407.298ms
benchCombined*10000: 2736.522ms
```

**After sync writing**

```
benchString*10000: 628.928ms
benchError*10000: 713.008ms
benchCombined*10000: 959.710ms
```

**After async writing**
```
benchString*10000: 277.393ms
benchError*10000: 381.088ms
benchCombined*10000: 559.997ms
```

Benchmarks for every incremental change with `noop` writing

**Baseline**

```
benchString*10000: 1718.167ms
benchError*10000: 1806.478ms
benchCombined*10000: 2112.879ms
```

**Replace json-stringify-safe  with fast-safe-stringify**

```
benchString*10000: 1497.956ms
benchError*10000: 1491.109ms
benchCombined*10000: 1814.321ms
```

**Replace Hoek.applyToDefaults with Object.assign**

```
benchString*10000: 784.983ms
benchError*10000: 785.215ms
benchCombined*10000: 1020.040ms
```

**Remove moment timestamp formatting**

```
benchString*10000: 380.328ms
benchError*10000: 498.167ms
benchCombined*10000: 664.189ms
```

**Console.log vs Sonic boom for writing**

**Async**

```
benchSonic*10000: 106.746ms
benchConsole*10000: 461.731ms
```

**Sync**

```
benchSonic*10000: 329.432ms
benchConsole*10000: 450.533ms
```