@aptomaKetil 2020-04-17
Does it default to sync or async? I see the default option passed to the sonic constructor is `sync: true`, but the flush method seems to default to async.