@martinj 2020-04-17
> Does it default to sync or async? I see the default option passed to the sonic constructor is `sync: true`, but the flush method seems to default to async.

It defaults to sync. Flush method is only exposed since using async and not loosing logs would require process hooks to flush. Flush behaviour is not relevant to if you run async or not for writes.